+++ source/ftrack_api/event/hub.py
@@ -14,6 +14,7 @@ import operator
 import functools
 import json
 import socket
+import warnings
 
 import requests
 import requests.exceptions
@@ -40,8 +41,19 @@ ServerDetails = collections.namedtuple('ServerDetails', [
 ])
 
 
+
+
 class EventHub(object):
     '''Manage routing of events.'''
+
+    _future_signature_warning = (
+        'When constructing your Session object you did not explicitly define '
+        'auto_connect_event_hub as True even though you appear to be publishing '
+        'and / or subscribing to asynchronous events. In version version 2.0 of '
+        'the ftrack-python-api the default behavior will change from True '
+        'to False. Please make sure to update your tools. You can read more at '
+        'http://ftrack-python-api.rtd.ftrack.com/en/stable/release/migration.html'
+    )
 
     def __init__(self, server_url, api_user, api_key):
         '''Initialise hub, connecting to ftrack *server_url*.
@@ -75,6 +87,8 @@ class EventHub(object):
         # disconnection. Equates to 5 minutes.
         self._auto_reconnect_attempts = 30
         self._auto_reconnect_delay = 10
+
+        self._deprecation_warning_auto_connect = False
 
         # Mapping of Socket.IO codes to meaning.
         self._code_name_mapping = {
@@ -134,6 +148,9 @@ class EventHub(object):
         connected or connection fails.
 
         '''
+
+        self._deprecation_warning_auto_connect = False
+
         if self.connected:
             raise ftrack_api.exception.EventHubConnectionError(
                 'Already connected.'
@@ -543,6 +560,11 @@ class EventHub(object):
         event will be caught by this method and ignored.
 
         '''
+        if self._deprecation_warning_auto_connect and not synchronous:
+            warnings.warn(
+                self._future_signature_warning, FutureWarning
+            )
+
         try:
             return self._publish(
                 event, synchronous=synchronous, on_reply=on_reply
@@ -700,18 +722,23 @@ class EventHub(object):
 
             # Automatically publish a non None response as a reply when not in
             # synchronous mode.
-            if not synchronous and response is not None:
-
-                try:
-                    self.publish_reply(
-                        event, data=response, source=subscriber.metadata
+            if not synchronous:
+                if self._deprecation_warning_auto_connect:
+                    warnings.warn(
+                        self._future_signature_warning, FutureWarning
                     )
 
-                except Exception:
-                    self.logger.exception(L(
-                        'Error publishing response {0} from subscriber {1} '
-                        'for event {2}.', response, subscriber, event
-                    ))
+                if response is not None:
+                    try:
+                        self.publish_reply(
+                            event, data=response, source=subscriber.metadata
+                        )
+
+                    except Exception:
+                        self.logger.exception(L(
+                            'Error publishing response {0} from subscriber {1} '
+                            'for event {2}.', response, subscriber, event
+                        ))
 
             # Check whether to continue processing topic event.
             if event.is_stopped():