+++ source/ftrack_api/session.py
@@ -69,7 +69,7 @@ class Session(object):
     def __init__(
         self, server_url=None, api_key=None, api_user=None, auto_populate=True,
         plugin_paths=None, cache=None, cache_key_maker=None,
-        auto_connect_event_hub=True, schema_cache_path=None,
+        auto_connect_event_hub=None, schema_cache_path=None,
         plugin_arguments=None
     ):
         '''Initialise session.
@@ -230,10 +230,11 @@ class Session(object):
         self._event_hub = ftrack_api.event.hub.EventHub(
             self._server_url,
             self._api_user,
-            self._api_key
+            self._api_key,
         )
 
-        if auto_connect_event_hub:
+
+        if auto_connect_event_hub in (None, True):
             # Connect to event hub in background thread so as not to block main
             # session usage waiting for event hub connection.
             self._auto_connect_event_hub_thread = threading.Thread(
@@ -241,6 +242,12 @@ class Session(object):
             )
             self._auto_connect_event_hub_thread.daemon = True
             self._auto_connect_event_hub_thread.start()
+
+        # To help with migration from auto_connect_event_hub default changing
+        # from True to False.
+        self._event_hub._deprecation_warning_auto_connect = (
+            auto_connect_event_hub is None
+        )
 
         # Register to auto-close session on exit.
         atexit.register(self.close)