+++ source/ftrack_api/session.py
@@ -280,7 +280,7 @@ class Session(object):
         self._configure_locations()
         self.event_hub.publish(
             ftrack_api.event.base.Event(
-                topic='ftrack.api.session.initialized',
+                topic='ftrack.api.session.ready',
                 data=dict(
                     session=self
                 )
@@ -443,6 +443,16 @@ class Session(object):
 
         # Re-configure certain session aspects that may be dependant on cache.
         self._configure_locations()
+
+        self.event_hub.publish(
+            ftrack_api.event.base.Event(
+                topic='ftrack.api.session.reset',
+                data=dict(
+                    session=self
+                )
+            ),
+            synchronous=True
+        )
 
     def auto_populating(self, auto_populate):
         '''Temporarily set auto populate to *auto_populate*.