+++ doc/release/release_notes.rst
@@ -11,6 +11,19 @@ Release Notes
 
 .. release:: Upcoming
 
+    .. change:: new
+        :tags: depreciation, events
+
+        In version 2.0.0 of the `ftrack-python-api` the default behavior for
+        the `ftrack_api.Session` class will change for the argument
+        `auto_connect_event_hub`, the default value will switch from True to False.
+
+        A warning will now be emitted if async events are published or
+        subscribed to without auto_connect_event_hub has not explicitly been
+        set to True.
+
+        .. seealso:: :ref:`release/migration/2.0.0/event_hub`.
+
     .. change:: fixed
         :tags: documentation